feat: settings lazy loading
