+++ apps/Conduit-UI/src/pages/settings/password.tsx
@@ -1,6 +1,13 @@
 import React, { ReactElement } from 'react';
 import SettingsLayout from '../../components/navigation/InnerLayouts/settingsLayout';
-import ChangePasswordTab from '../../components/settings/ChangePasswordTab';
+import dynamic from 'next/dynamic';
+import ScaleLoader from 'react-spinners/ScaleLoader';
+
+const ChangePasswordTab = dynamic(() => import('../../components/settings/ChangePasswordTab'), {
+  loading: () => (
+    <ScaleLoader speedMultiplier={3} color={'#07D9C4'} loading={true} height={21} width={4} />
+  ),
+});
 
 const Password = () => {
   return <ChangePasswordTab />;
