+++ apps/Conduit-UI/src/pages/settings/settings.tsx
@@ -1,8 +1,15 @@
 import React, { ReactElement, useEffect } from 'react';
 import SettingsLayout from '../../components/navigation/InnerLayouts/settingsLayout';
-import CoreSettingsTab from '../../components/settings/CoreSettingsTab';
 import { asyncGetAdminSettings, asyncGetCoreSettings } from '../../redux/slices/settingsSlice';
 import { useAppDispatch } from '../../redux/store';
+import dynamic from 'next/dynamic';
+import ScaleLoader from 'react-spinners/ScaleLoader';
+
+const CoreSettingsTab = dynamic(() => import('../../components/settings/CoreSettingsTab'), {
+  loading: () => (
+    <ScaleLoader speedMultiplier={3} color={'#07D9C4'} loading={true} height={21} width={4} />
+  ),
+});
 
 const CoreSettings = () => {
   const dispatch = useAppDispatch();
